casetest: split test for improving ci
